+++ test/python/tools/jupyter/test_notebooks.py
@@ -13,7 +13,7 @@ import unittest
 import nbformat
 from nbconvert.preprocessors import ExecutePreprocessor
 from qiskit.tools.visualization._matplotlib import HAS_MATPLOTLIB
-from qiskit.test import (Path, QiskitTestCase, requires_qe_access, slow_test)
+from qiskit.test import (Path, QiskitTestCase, requires_qe_access)
 
 
 # Timeout (in seconds) for a single notebook.
@@ -58,7 +58,6 @@ class TestJupyter(QiskitTestCase):
 
     @unittest.skipIf(not HAS_MATPLOTLIB, 'matplotlib not available.')
     @requires_qe_access
-    @slow_test
     def test_backend_tools(self, qe_token, qe_url):
         """Test Jupyter backend tools."""
         self._execute_notebook(self._get_resource_path(